requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
httpx==0.27.2
fal-client==0.5.4
boto3==1.35.0
